file editors - move the input to the editor (`browser`)
