+++ src/vs/workbench/contrib/files/electron-sandbox/textFileEditor.ts
@@ -5,7 +5,7 @@
 
 import { localize } from 'vs/nls';
 import { TextFileEditor } from 'vs/workbench/contrib/files/browser/editors/textFileEditor';
-import { FileEditorInput } from 'vs/workbench/contrib/files/common/editors/fileEditorInput';
+import { FileEditorInput } from 'vs/workbench/contrib/files/browser/editors/fileEditorInput';
 import { EditorOptions } from 'vs/workbench/common/editor';
 import { FileOperationError, FileOperationResult, IFileService, MIN_MAX_MEMORY_SIZE_MB, FALLBACK_MAX_MEMORY_SIZE_MB } from 'vs/platform/files/common/files';
 import { createErrorWithActions } from 'vs/base/common/errors';
